+++ Lab2/imageFunctions.h
@@ -2,6 +2,6 @@
 
 #ifndef IMAGEFUNCTIONS_H
 #define IMAGEFUNCTIONS_H
-uint16_t EightBitHistogram_C(uint16_t width, uint16_t height, const uint8_t *p_image, uint8_t *p_histogram);
+uint16_t EightBitHistogram_C(uint16_t width, uint16_t height, const uint8_t *p_image, uint16_t *p_histogram);
 
 #endif
